Change set summary:
 1) BugFix: @ViewWith overrides on methods now work if a forClass is not defined
 2) Updated the POM to reference latest versions of Resteasy
 3) Removed the older ResteasyFilter
 4) Added a new HtmleasyFilterDispatcher (and HtmleasyServletDispatcher for completeness)
 5) Added an HtmleasyProviders utility class.
 6) Added a Path utility class to assist with resolving paths to 
    controllers/methods (type-safe/refactor-save controller path references)
 7) Added extra constructors to RedirectException() so it's possible to 
    redirect to the Class or it's method (same as Path.to())
 8) Added a @ViewSet annotation allowing multiple @ViewWith decelerations to 
    be defined on a method.  e.g. Select the best view of the returned 
    model/class type.

Code Review: Changes reviewed by Jeff pre-commit.

Detail/Rational:

Change 1) File: providers/ViewResolver.java - It was not possible to define a ViewWith annotation on a method without a forClass value.  This was contrary to the current documentation.

Change 4) These files replace the existing RestEasyFilter.  They work by delaying the open/fetch of the output stream until there is an actual write.  Its simpler than the current filter and inherits the logic provided by the inbuilt RestEasy filter. It also should be easier to maintain over time.

Change 5) Classpath scanning to register @Providers is slow. On Google App Engine for example, it's better to define an explicit Application class and provide a list.  This utility class provides a list of the Provider classes to register avoiding the need for the user to know what provider classes the plugin provides.  e.g.

    public class Application extends javax.ws.rs.core.Application {

        @Override
        public Set<Class<?>> getClasses() {
            Set<Class<?>> classes = new HashSet<Class<?>>();
            classes.add(MyRestStuff.class);
            classes.addAll(HtmleasyProviders.getClasses());
            return classes;
        }
    }


Change 6) This takes an idea form Microsoft's MVC3.  It's a simple wrapper around some JAX-RS classes and allows you to determine the path to a controller by referencing the controller class (e.g. resolve a route/path form a controller class or method).  This allows you to write code like:

    <form method="POST" action="<%= Path.to(MyController.class) %>">

or

    model.setPostUrl(Path.to(MyController.class));

Change 7) A similar concept was added to RedirectException (i.e. redirect to a path corresponding to a controller class/method).

Change 8) The documentation seemed to imply that it was possible to add "one or more @ViewWith annotations" however to my knowlage it's not possible to duplicate an annotation?  I worked around this with a @ViewSet annotation that may contain a list of @ViewWith items. Changes in ViewResolver.java. Wiki documentation needs updating.